debug: Add render logs to ALL wrapper components
Added console.log at the TOP of EVERY component render:
- ProfileSetupClient:
- RequireAuthClientOnly:
- RequireAuth:
- RequireIncompleteProfile:
- ProfileSetupForm:

This will show WHICH component is mounting and WHICH is blocking.
If we see  but not , the problem is in RequireAuthClientOnly.
If we see  but not , the problem is in RequireAuth.
If we see  but not , the problem is in RequireIncompleteProfile.
If we see  but not , the problem is RequireIncompleteProfile blocking.

This will definitively show WHERE the render chain breaks.